backend/requirements.txt
cloudinary==1.44.0
fastapi==0.115.12
numpy
pillow==11.2.1
pydantic==2.11.4
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
uvicorn==0.34.2
mlflow==2.8.1
prometheus-client
